feat: Add a way to limit the number of results from search_files tool
Currently search_files will produce up to 300 results (an arbitrary
number). In cortain cases this is too large number because only first
few occurrences are wanted. Unfortunately if too generic search pattern
is selected, lots of results will be returned to the LLM and this will
reduce remaining context significantly and require more time to process
than needed.

To address this, an explicit optional max_results parameter has been
added to the search_files tool. Its explanation is simple and it does
not increase the Roo Code prompt size too much.

If the max_results parameter is not supplied, code works exactly as
before.