✨ feat(extension): add auto-reload for core changes in dev mode
Previously, devs would need to restart the extension host to see core changes.
Now, with this change, we monitor the files within the core directory.
If any changes are detected, we trigger a reload command within the extension, effectively enabling auto-reloading within the extension

Dev notes
I was originally going to make this a new env var something like ENABLE_CORE_AUTO_RESTART, but instead just based it on the NODE_ENV. Let me know if you think the env var would be better! (or something else!)

- enhance development experience by eliminating manual restarts
- implement file system watcher for core files
- automatically reload extension host on changes

Test plan: Verified running the extension locally that changes to `src` reload the window in dev.