fix: improve i18n key detection in source code
Added a new pattern to match i18n keys used with parameters, which fixes
detection of keys like common:errors.invalid_mcp_settings_validation.

Refactored the code to split it into 4 functions:
- accumulateSourceKeys(): Collects all keys used in source code
- accumulateTranslationKeys(): Collects all keys found in translation files
- getKeysInSourceNotInTranslation(): Returns keys in source not in translations
- getKeysInTranslationNotInSource(): Returns keys in translations not in source

Added tests to validate that there are no keys in source not in translations
and no keys in translations not in source.

Signed-off-by: Eric Wheeler <roo-code@z.ewheeler.org>