Expands non-interactive click detection to include VSCode elements
Enhances the non-interactive click listener by adding support for 'VSCODE-TEXT-AREA' and 'VSCODE-TEXT-FIELD' tags. This ensures that typing in VSCode-specific input elements is not disrupted by unintentional triggers.

Not including these element was causing users not to be able to edit some input boxes on the settings page. This fixes that!