refactor: make updateApiReqMsg transactional
The updateApiReqMsg function was modifying the clineMessages array
directly and relying on a subsequent, separate call to
modifyClineMessages to persist the change. This was not atomic.

This change refactors updateApiReqMsg to be an async function that
encapsulates the entire transaction of updating the API request
message and saving it to disk within a single modifyClineMessages
call. The separate, redundant calls to modifyClineMessages have been
removed.

Signed-off-by: Eric Wheeler <roo-code@z.ewheeler.org>