refactor: make recursivelyMakeClineRequests message updates atomic
Refactors the `recursivelyMakeClineRequests` method to ensure message updates are transactional.

This change replaces a direct mutation of the `clineMessages` array followed by a separate save operation with a single call to `modifyConversation`. By doing so, it guarantees that the modification of the `api_req_started` message and the addition of the user's content to the conversation history occur as a single, atomic operation, preventing potential data inconsistencies.

Signed-off-by: Eric Wheeler <roo-code@z.ewheeler.org>